requirements.txt
APScheduler~=3.11.0
numpy~=2.3.2
okx~=2.1.2
okx_sdk~=5.5.812
pandas~=2.1.4+dfsg
pybit~=5.11.0
pymongo~=4.14.1
python_binance~=1.0.27
redis~=6.4.0
Requests~=2.32.5
